wip: PWM restructure
