+++ embassy-rp/src/pwm/mod.rs
@@ -1,0 +1,568 @@
+//! Pulse Width Modulation (PWM)
+
+use core::ops::Div;
+
+use embassy_hal_internal::{into_ref, Peripheral, PeripheralRef};
+use embedded_hal_1::pwm::ErrorKind;
+use fixed::traits::ToFixed;
+use fixed::FixedU16;
+use pac::pwm::regs::{ChDiv, Intr};
+use pac::pwm::vals::Divmode;
+
+use crate::clocks::clk_sys_freq;
+use crate::gpio::{AnyPin, Pin as GpioPin, Pull, SealedPin as _};
+use crate::{pac, peripherals, RegExt};
+
+use self::v2::Channel;
+
+pub mod builder;
+pub mod v2;
+
+pub const MIN_PWM_FREQ: f32 = 7.5;
+
+/// The configuration of a PWM slice.
+/// Note the period in clock cycles of a slice can be computed as:
+/// `(top + 1) * (phase_correct ? 1 : 2) * divider`
+#[non_exhaustive]
+#[derive(Clone)]
+pub struct Config {
+    /// Inverts the PWM output signal on channel A.
+    pub invert_a: bool,
+    /// Inverts the PWM output signal on channel B.
+    pub invert_b: bool,
+    /// Enables phase-correct mode for PWM operation.
+    /// In phase-correct mode, the PWM signal is generated in such a way that
+    /// the pulse is always centered regardless of the duty cycle.
+    /// The output frequency is halved when phase-correct mode is enabled.
+    pub phase_correct: bool,
+    /// Enables the PWM slice, allowing it to generate an output.
+    pub enable: bool,
+    /// A fractional clock divider, represented as a fixed-point number with
+    /// 8 integer bits and 4 fractional bits. It allows precise control over
+    /// the PWM output frequency by gating the PWM counter increment.
+    /// A higher value will result in a slower output frequency.
+    pub divider: fixed::FixedU16<fixed::types::extra::U4>,
+    /// The output on channel A goes high when `compare_a` is higher than the
+    /// counter. A compare of 0 will produce an always low output, while a
+    /// compare of `top + 1` will produce an always high output.
+    pub compare_a: u16,
+    /// The output on channel B goes high when `compare_b` is higher than the
+    /// counter. A compare of 0 will produce an always low output, while a
+    /// compare of `top + 1` will produce an always high output.
+    pub compare_b: u16,
+    /// The point at which the counter wraps, representing the maximum possible
+    /// period. The counter will either wrap to 0 or reverse depending on the
+    /// setting of `phase_correct`.
+    pub top: u16,
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        Self {
+            invert_a: false,
+            invert_b: false,
+            phase_correct: false,
+            enable: true, // differs from reset value
+            divider: 1.to_fixed(),
+            compare_a: 0,
+            compare_b: 0,
+            top: 0xffff,
+        }
+    }
+}
+
+/// PWM input mode.
+pub enum InputMode {
+    /// The fractional divider operation is gated by the PWM B pin.
+    Level,
+    /// The counter advances with each rising edge of the PWM B pin.
+    RisingEdge,
+    /// The counter advances with each falling edge of the PWM B pin.
+    FallingEdge,
+}
+
+impl From<InputMode> for Divmode {
+    fn from(value: InputMode) -> Self {
+        match value {
+            InputMode::Level => Divmode::LEVEL,
+            InputMode::RisingEdge => Divmode::RISE,
+            InputMode::FallingEdge => Divmode::FALL,
+        }
+    }
+}
+
+/// PWM driver.
+pub struct Pwm<'d, T: Slice> {
+    inner: PeripheralRef<'d, T>,
+    pin_a: Option<PeripheralRef<'d, AnyPin>>,
+    pin_b: Option<PeripheralRef<'d, AnyPin>>,
+    freq: u32,
+    duty_a: Option<u16>,
+    duty_b: Option<u16>,
+}
+
+impl<'d, T: Slice> Pwm<'d, T> {
+    fn new_inner(
+        inner: impl Peripheral<P = T> + 'd,
+        a: Option<PeripheralRef<'d, AnyPin>>,
+        b: Option<PeripheralRef<'d, AnyPin>>,
+        b_pull: Pull,
+        config: Config,
+        divmode: Divmode,
+    ) -> Self {
+        into_ref!(inner);
+
+        let p = inner.regs();
+        p.csr().modify(|w| {
+            w.set_divmode(divmode);
+            w.set_en(false);
+        });
+        p.ctr().write(|w| w.0 = 0);
+        Self::configure(p, &config);
+
+        if let Some(pin) = &a {
+            pin.gpio().ctrl().write(|w| w.set_funcsel(4));
+        }
+        if let Some(pin) = &b {
+            pin.gpio().ctrl().write(|w| w.set_funcsel(4));
+            pin.pad_ctrl().modify(|w| {
+                w.set_pue(b_pull == Pull::Up);
+                w.set_pde(b_pull == Pull::Down);
+            });
+        }
+        Self {
+            inner,
+            pin_a: a,
+            pin_b: b,
+            freq: clk_sys_freq(),
+            duty_a: None,
+            duty_b: None,
+        }
+    }
+
+    /// Create PWM driver without any configured pins.
+    #[inline]
+    pub fn new_free(inner: impl Peripheral<P = T> + 'd, config: Config) -> Self {
+        Self::new_inner(inner, None, None, Pull::None, config, Divmode::DIV)
+    }
+
+    /// Create PWM driver with a single 'a' as output.
+    #[inline]
+    pub fn new_output_a(
+        inner: impl Peripheral<P = T> + 'd,
+        a: impl Peripheral<P = impl ChannelAPin<T>> + 'd,
+        config: Config,
+    ) -> Self {
+        into_ref!(a);
+        Self::new_inner(inner, Some(a.map_into()), None, Pull::None, config, Divmode::DIV)
+    }
+
+    /// Create PWM driver with a single 'b' pin as output.
+    #[inline]
+    pub fn new_output_b(
+        inner: impl Peripheral<P = T> + 'd,
+        b: impl Peripheral<P = impl ChannelBPin<T>> + 'd,
+        config: Config,
+    ) -> Self {
+        into_ref!(b);
+        Self::new_inner(inner, None, Some(b.map_into()), Pull::None, config, Divmode::DIV)
+    }
+
+    /// Create PWM driver with a 'a' and 'b' pins as output.
+    #[inline]
+    pub fn new_output_ab(
+        inner: impl Peripheral<P = T> + 'd,
+        a: impl Peripheral<P = impl ChannelAPin<T>> + 'd,
+        b: impl Peripheral<P = impl ChannelBPin<T>> + 'd,
+        config: Config,
+    ) -> Self {
+        into_ref!(a, b);
+        Self::new_inner(
+            inner,
+            Some(a.map_into()),
+            Some(b.map_into()),
+            Pull::None,
+            config,
+            Divmode::DIV,
+        )
+    }
+
+    /// Create PWM driver with a single 'b' as input pin.
+    #[inline]
+    pub fn new_input(
+        inner: impl Peripheral<P = T> + 'd,
+        b: impl Peripheral<P = impl ChannelBPin<T>> + 'd,
+        b_pull: Pull,
+        mode: InputMode,
+        config: Config,
+    ) -> Self {
+        into_ref!(b);
+        Self::new_inner(inner, None, Some(b.map_into()), b_pull, config, mode.into())
+    }
+
+    /// Create PWM driver with a 'a' and 'b' pins in the desired input mode.
+    #[inline]
+    pub fn new_output_input(
+        inner: impl Peripheral<P = T> + 'd,
+        a: impl Peripheral<P = impl ChannelAPin<T>> + 'd,
+        b: impl Peripheral<P = impl ChannelBPin<T>> + 'd,
+        b_pull: Pull,
+        mode: InputMode,
+        config: Config,
+    ) -> Self {
+        into_ref!(a, b);
+        Self::new_inner(
+            inner,
+            Some(a.map_into()),
+            Some(b.map_into()),
+            b_pull,
+            config,
+            mode.into(),
+        )
+    }
+
+    /// Set the PWM config.
+    #[inline]
+    pub fn set_config(&mut self, config: &Config) {
+        Self::configure(self.inner.regs(), config);
+    }
+
+    #[inline]
+    fn configure(p: pac::pwm::Channel, config: &Config) {
+        if config.divider > FixedU16::<fixed::types::extra::U4>::from_bits(0xFFF) {
+            panic!("Requested divider is too large");
+        }
+
+        p.div().write_value(ChDiv(config.divider.to_bits() as u32));
+        p.cc().write(|w| {
+            w.set_a(config.compare_a);
+            w.set_b(config.compare_b);
+        });
+        p.top().write(|w| w.set_top(config.top));
+        p.csr().modify(|w| {
+            w.set_a_inv(config.invert_a);
+            w.set_b_inv(config.invert_b);
+            w.set_ph_correct(config.phase_correct);
+            w.set_en(config.enable);
+        });
+    }
+
+    /// Sets the PWM frequency for this slice. Note that this affects
+    /// both channels (A and B).
+    ///
+    /// @param freq: The frequency in Hz.
+    ///
+    /// TODO: Add support for fractional hz
+    #[inline]
+    pub fn set_freq(&mut self, freq: u32) -> Result<(), ErrorKind> {
+        let clock_hz = clk_sys_freq() / 1_000_000;
+        let min_hz = clock_hz / 125;
+        let max_hz = clock_hz / 2;
+
+        // The frequency must be between 8 and the clock frequency
+        if freq < MIN_PWM_FREQ as u32 || freq > max_hz as u32 {
+            return Err(ErrorKind::Other);
+        }
+
+        // Only perform recalculations if the frequency has changed
+        if freq == self.freq {
+            return Ok(());
+        }
+
+        self.freq = freq;
+
+        self.recalculate_div_wrap(freq)?;
+
+        if let Some(duty_a) = self.duty_a {
+            self.recalculate_duty(Channel::A, duty_a)?;
+        }
+
+        if let Some(duty_b) = self.duty_b {
+            self.recalculate_duty(Channel::B, duty_b)?;
+        }
+        Ok(())
+    }
+
+    /// Get the current PWM frequency for this slice.
+    #[inline]
+    pub fn freq(&self) -> u32 {
+        self.freq
+    }
+
+    /// Set the PWM duty cycle for channel A. Returns an error if the slice
+    /// does not have a channel A configured.
+    #[inline]
+    pub fn set_duty_a(&mut self, duty: u16) -> Result<(), ErrorKind> {
+        if self.pin_a.is_none() {
+            return Err(ErrorKind::Other);
+        }
+
+        self.recalculate_duty(Channel::A, duty)?;
+        self.duty_a = Some(duty);
+        Ok(())
+    }
+
+    /// Set the PWM duty cycle for channel B. Returns an error if the slice
+    /// does not have a channel B configured.
+    #[inline]
+    pub fn set_duty_b(&mut self, duty: u16) -> Result<(), ErrorKind> {
+        if self.pin_b.is_none() {
+            return Err(ErrorKind::Other);
+        }
+
+        self.recalculate_duty(Channel::B, duty)?;
+        self.duty_b = Some(duty);
+        Ok(())
+    }
+
+    /// Set the PWM duty cycle for both channels A and B. Returns an error if
+    /// the slice does not have both A & B channels configured.
+    #[inline]
+    pub fn set_duty_ab(&mut self, duty: u16) -> Result<(), ErrorKind> {
+        if self.pin_a.is_none() || self.pin_b.is_none() {
+            return Err(ErrorKind::Other);
+        }
+
+        self.set_duty_a(duty)?;
+        self.set_duty_b(duty)
+    }
+
+    /// Advances a slice’s output phase by one count while it is running
+    /// by inserting a pulse into the clock enable. The counter
+    /// will not count faster than once per cycle.
+    #[inline]
+    pub fn phase_advance(&mut self) {
+        let p = self.inner.regs();
+        p.csr().write_set(|w| w.set_ph_adv(true));
+        while p.csr().read().ph_adv() {}
+    }
+
+    /// Retards a slice’s output phase by one count while it is running
+    /// by deleting a pulse from the clock enable. The counter will not
+    /// count backward when clock enable is permenantly low.
+    #[inline]
+    pub fn phase_retard(&mut self) {
+        let p = self.inner.regs();
+        p.csr().write_set(|w| w.set_ph_ret(true));
+        while p.csr().read().ph_ret() {}
+    }
+
+    /// Read PWM counter.
+    #[inline]
+    pub fn counter(&self) -> u16 {
+        self.inner.regs().ctr().read().ctr()
+    }
+
+    /// Write PWM counter.
+    #[inline]
+    pub fn set_counter(&self, ctr: u16) {
+        self.inner.regs().ctr().write(|w| w.set_ctr(ctr))
+    }
+
+    /// Wait for channel interrupt.
+    #[inline]
+    pub fn wait_for_wrap(&mut self) {
+        while !self.wrapped() {}
+        self.clear_wrapped();
+    }
+
+    /// Check if interrupt for channel is set.
+    #[inline]
+    pub fn wrapped(&mut self) -> bool {
+        pac::PWM.intr().read().0 & self.bit() != 0
+    }
+
+    #[inline]
+    /// Clear interrupt flag.
+    pub fn clear_wrapped(&mut self) {
+        pac::PWM.intr().write_value(Intr(self.bit() as _));
+    }
+
+    /// Enables the PWM counter for this slice. Note that enablement can only
+    /// be specified on the slice-level, so this affects both PWM channels
+    /// (A & B).
+    #[inline]
+    pub fn enable(&mut self) {
+        self.inner.regs().csr().write(|w| w.set_en(true));
+    }
+
+    /// Disables the PWM counter for this slice. Note that enablement can only
+    /// be specified on the slice-level, so this affects both PWM channels
+    /// (A & B).
+    #[inline]
+    pub fn disable(&mut self) {
+        self.inner.regs().csr().write(|w| w.set_en(false));
+    }
+
+    #[inline]
+    fn bit(&self) -> u32 {
+        1 << self.inner.number() as usize
+    }
+
+    /// Recalculates the TOP and DIV values and updates the PWM slice registers.
+    #[inline]
+    fn recalculate_div_wrap(&self, freq: u32) -> Result<(), ErrorKind> {
+        let mut clk_divider = 0;
+        let mut wrap = 0;
+        let mut clock_div;
+        let clock = clk_sys_freq();
+
+        // Only recalculate divider if frequency has changed
+        for div in 1..u8::MAX as u32 {
+            clk_divider = div;
+            // Find clock_division to fit current frequency
+            clock_div = clock.div(div);
+            wrap = clock_div / freq;
+            if clock_div / u16::MAX as u32 <= freq && wrap <= u16::MAX as u32 {
+                break;
+            }
+        }
+
+        if clk_divider < u8::MAX as u32 {
+            // Only update divider and top registers if the frequency has
+            // changed
+            self.inner.regs().div().write(|w| w.set_int(clk_divider as u8));
+            self.inner.regs().top().write(|w| w.set_top(wrap as u16));
+            Ok(())
+        } else {
+            Err(ErrorKind::Other)
+        }
+    }
+
+    /// Recalculates the duty cycle for a channel and updates the PWM slice CC
+    /// register with the new values for both A and B channels.
+    #[inline]
+    fn recalculate_duty(&self, channel: Channel, duty: u16) -> Result<(), ErrorKind> {
+        // Get the current `DIV` register value for this slice (only the
+        // integer part is used)
+        let wrap = self.inner.regs().div().read().int();
+
+        // Update the `CC` register.
+        self.inner.regs().cc().write(|w| {
+            let compare = ((((wrap + if duty == 100 { 1 } else { 0 }) as u16) * duty) / 100) as u16;
+
+            if self.pin_a.is_some() && channel == Channel::A {
+                w.set_a(compare);
+            } else if self.pin_b.is_some() && channel == Channel::B {
+                w.set_b(compare);
+            } else {
+                // We shouldn't have been able to get here if the `Pwm`
+                // constructors are doing their job.
+                return Err(ErrorKind::Other);
+            }
+            Ok(())
+        })
+    }
+}
+
+/// Batch representation of PWM slices.
+pub struct PwmBatch(u32);
+
+impl PwmBatch {
+    #[inline]
+    /// Enable a PWM slice in this batch.
+    pub fn enable(&mut self, pwm: &Pwm<'_, impl Slice>) {
+        self.0 |= pwm.bit();
+    }
+
+    #[inline]
+    /// Enable slices in this batch in a PWM.
+    pub fn set_enabled(enabled: bool, batch: impl FnOnce(&mut PwmBatch)) {
+        let mut en = PwmBatch(0);
+        batch(&mut en);
+        if enabled {
+            pac::PWM.en().write_set(|w| w.0 = en.0);
+        } else {
+            pac::PWM.en().write_clear(|w| w.0 = en.0);
+        }
+    }
+}
+
+impl<'d, T: Slice> Drop for Pwm<'d, T> {
+    fn drop(&mut self) {
+        self.inner.regs().csr().write_clear(|w| w.set_en(false));
+        if let Some(pin) = &self.pin_a {
+            pin.gpio().ctrl().write(|w| w.set_funcsel(31));
+        }
+        if let Some(pin) = &self.pin_b {
+            pin.gpio().ctrl().write(|w| w.set_funcsel(31));
+        }
+    }
+}
+
+trait SealedSlice {}
+
+/// PWM Slice.
+#[allow(private_bounds)]
+pub trait Slice: Peripheral<P = Self> + SealedSlice + Sized + 'static {
+    /// Slice number.
+    fn number(&self) -> u8;
+
+    /// Slice register block.
+    fn regs(&self) -> pac::pwm::Channel {
+        pac::PWM.ch(self.number() as _)
+    }
+}
+
+macro_rules! slice {
+    ($name:ident, $num:expr) => {
+        impl SealedSlice for peripherals::$name {}
+        impl Slice for peripherals::$name {
+            fn number(&self) -> u8 {
+                $num
+            }
+        }
+        //impl SealedSlice for PwmFreeRunningSlice<'_, peripherals::$name> {}
+    };
+}
+
+slice!(PWM_SLICE0, 0);
+slice!(PWM_SLICE1, 1);
+slice!(PWM_SLICE2, 2);
+slice!(PWM_SLICE3, 3);
+slice!(PWM_SLICE4, 4);
+slice!(PWM_SLICE5, 5);
+slice!(PWM_SLICE6, 6);
+slice!(PWM_SLICE7, 7);
+
+/// PWM Channel A.
+pub trait ChannelAPin<T: Slice>: GpioPin {}
+/// PWM Channel B.
+pub trait ChannelBPin<T: Slice>: GpioPin {}
+
+macro_rules! impl_pin {
+    ($pin:ident, $channel:ident, $kind:ident) => {
+        impl $kind<peripherals::$channel> for peripherals::$pin {}
+    };
+}
+
+impl_pin!(PIN_0, PWM_SLICE0, ChannelAPin);
+impl_pin!(PIN_1, PWM_SLICE0, ChannelBPin);
+impl_pin!(PIN_2, PWM_SLICE1, ChannelAPin);
+impl_pin!(PIN_3, PWM_SLICE1, ChannelBPin);
+impl_pin!(PIN_4, PWM_SLICE2, ChannelAPin);
+impl_pin!(PIN_5, PWM_SLICE2, ChannelBPin);
+impl_pin!(PIN_6, PWM_SLICE3, ChannelAPin);
+impl_pin!(PIN_7, PWM_SLICE3, ChannelBPin);
+impl_pin!(PIN_8, PWM_SLICE4, ChannelAPin);
+impl_pin!(PIN_9, PWM_SLICE4, ChannelBPin);
+impl_pin!(PIN_10, PWM_SLICE5, ChannelAPin);
+impl_pin!(PIN_11, PWM_SLICE5, ChannelBPin);
+impl_pin!(PIN_12, PWM_SLICE6, ChannelAPin);
+impl_pin!(PIN_13, PWM_SLICE6, ChannelBPin);
+impl_pin!(PIN_14, PWM_SLICE7, ChannelAPin);
+impl_pin!(PIN_15, PWM_SLICE7, ChannelBPin);
+impl_pin!(PIN_16, PWM_SLICE0, ChannelAPin);
+impl_pin!(PIN_17, PWM_SLICE0, ChannelBPin);
+impl_pin!(PIN_18, PWM_SLICE1, ChannelAPin);
+impl_pin!(PIN_19, PWM_SLICE1, ChannelBPin);
+impl_pin!(PIN_20, PWM_SLICE2, ChannelAPin);
+impl_pin!(PIN_21, PWM_SLICE2, ChannelBPin);
+impl_pin!(PIN_22, PWM_SLICE3, ChannelAPin);
+impl_pin!(PIN_23, PWM_SLICE3, ChannelBPin);
+impl_pin!(PIN_24, PWM_SLICE4, ChannelAPin);
+impl_pin!(PIN_25, PWM_SLICE4, ChannelBPin);
+impl_pin!(PIN_26, PWM_SLICE5, ChannelAPin);
+impl_pin!(PIN_27, PWM_SLICE5, ChannelBPin);
+impl_pin!(PIN_28, PWM_SLICE6, ChannelAPin);
+impl_pin!(PIN_29, PWM_SLICE6, ChannelBPin);
